Phase 3: split cross-workflow edit rendering
